+++ src/protocol/ratbox.c
@@ -899,8 +899,8 @@ class ProtoRatbox : public Module
 
 		moduleAddModes();
 
-		ircd->DefMLock[(size_t)CMODE_NOEXTERNAL] = true;
-		ircd->DefMLock[(size_t)CMODE_TOPIC] = true;
+		ircd->DefMLock[CMODE_NOEXTERNAL] = true;
+		ircd->DefMLock[CMODE_TOPIC] = true;
 
 		pmodule_ircd_proto(&ircd_proto);
 		moduleAddIRCDMsgs();
